Beta version 0.3.4! / Версия Beta 0.3.4! - changed resource prices; - changed the price calculation scheme; - other trade changes; - minor fixes; -------------- - изменил цены на ресурсы; - изменил схему вычисление цены; - прочие изменения торговли; - небольшие исправления;

+++ Assets/Scripts/Planets/Factory.cs
@@ -100,7 +100,7 @@ namespace PlanetsColony
             return _canLevelUp;
         }
 
-        private uint CalculateResourceValueMultiplier()
+        private float CalculateResourceValueMultiplier()
         {
             return _resourceValueMultiplier = _level;
         }
